+++ FE/src/ts/main.ts
@@ -1,6 +1,6 @@
-import TodoController from './todoController.js';
-import TodoView from './todoView.js';
-import TodoEventManager from './todoEventManager.js';
+import TodoController from './todoModules/todoController.js';
+import TodoView from './todoModules/todoView.js';
+import TodoEventManager from './todoModules/todoEventManager.js';
 
 const todoView = new TodoView();
 const todoEventManager = new TodoEventManager({ todoView });
